requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zipp==3.23.0
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
